Bug fixes for the video playback

+++ app/src/free/java/delit/piwigoclient/ui/slideshow/AlbumPictureItemFragment.java
@@ -3,7 +3,7 @@ package delit.piwigoclient.ui.slideshow;
 import delit.piwigoclient.model.piwigo.PictureResourceItem;
 
 public class AlbumPictureItemFragment extends AbstractAlbumPictureItemFragment {
-    public static AlbumPictureItemFragment newInstance(PictureResourceItem galleryItem, long albumResourceItemIdx, long albumResourceItemCount, long totalResourceItemCount) {
+    public static AlbumPictureItemFragment newInstance(PictureResourceItem galleryItem, int albumResourceItemIdx, int albumResourceItemCount, long totalResourceItemCount) {
         AlbumPictureItemFragment fragment = new AlbumPictureItemFragment();
         fragment.setArguments(buildArgs(galleryItem, albumResourceItemIdx, albumResourceItemCount, totalResourceItemCount));
         return fragment;
